+++ salesforce_functions/_internal/app.py
@@ -1,5 +1,7 @@
 import contextlib
+import os
 import sys
+from pathlib import Path
 from typing import Any, AsyncGenerator
 
 import orjson
@@ -13,10 +15,11 @@ from structlog.stdlib import BoundLogger
 from ..context import Context, Org, User
 from ..data_api import DataAPI
 from ..invocation_event import InvocationEvent
-from . import config
 from .cloud_event import CloudEventError, SalesforceFunctionsCloudEvent
 from .function_loader import LoadFunctionError, load_function
 from .logging import configure_logging, get_logger
+
+PROJECT_PATH_ENV_VAR = "FUNCTION_PROJECT_PATH"
 
 
 class OrjsonResponse(JSONResponse):
@@ -118,8 +121,11 @@ async def lifespan(app: Starlette) -> AsyncGenerator[None, None]:
     # time the function is invoked.
     app.state.logger = get_logger().bind()
 
+    # This env var is set by the CLI, as a way to propagate CLI args to the ASGI app.
+    project_path = Path(os.environ[PROJECT_PATH_ENV_VAR])
+
     try:
-        app.state.function = load_function(config.project_path())
+        app.state.function = load_function(project_path)
     except LoadFunctionError as e:
         # We cannot log an error message and `sys.exit(1)` like in the CLI's `check_function()`,
         # since we're running inside a uvicorn-managed coroutine. So instead, we raise an
